Standardise format of opening book on 32/64 bit Linux
The opening book was in different formats on 32 bit and 64 bit Linux
since it used long integers in the layout. long has been substituted
with int and the same book can be used on both systems.

+++ engine/book.h
@@ -57,9 +57,9 @@
 
 struct gdxadmin
 {
-    ULONG bookcount;
-    ULONG booksize;
-    ULONG maxoffset;
+    UINT bookcount;
+    UINT booksize;
+    UINT maxoffset;
 };
 
 extern struct gdxadmin B;
@@ -71,7 +71,7 @@ extern struct gdxadmin B;
 
 struct gdxdata
 {
-    ULONG  hashbd;
+    UINT   hashbd;
     USHORT hashkey;
     USHORT bmove;
     USHORT flags; /* flags BADMOVE, GOODMOVE,  LASTMOVE */
